Create tab component, where first content shows buttons, second content shows form input and third content shows material tree component
