+++ src/app/app.component.html
@@ -8,7 +8,7 @@
       <mat-icon>accessibility</mat-icon>
 
     </button>
-    <button mat-button>
+    <button mat-button routerLink="/tabs" routerLinkActive="active-link">
       <mat-icon>announcement</mat-icon>
 
     </button>
@@ -36,63 +36,6 @@
 
 <!-- <h3>Button</h3>
 
-<h3>Tabs</h3>
-<mat-tab-group>
-  <mat-tab label="First"> Content 1 </mat-tab>
-  <mat-tab label="Second"> Content 2 </mat-tab>
-  <mat-tab label="Third"> Content 3 </mat-tab>
-</mat-tab-group>
-
-<mat-tab-group mat-align-tabs="center" animationDuration="1000ms">
-  <mat-tab label="First">
-    <ng-template matTabContent>
-      The First Content
-    </ng-template>
-  </mat-tab>
-  <mat-tab label="Second">
-    <ng-template matTabContent>
-      The Second Content
-    </ng-template>
-  </mat-tab>
-  <mat-tab label="Third">
-    <ng-template matTabContent>
-      The Third Content
-    </ng-template>
-  </mat-tab>
-</mat-tab-group>
-
-<h3>Tree</h3>
-<mat-tree
-  [dataSource]="dataSource"
-  [treeControl]="treeControl"
-  class="example-tree"
->
-  <mat-tree-node *matTreeNodeDef="let node" matTreeNodeToggle>
-    <li class="mat-tree-node">
-      <button mat-icon-button disabled></button>
-      {{ node.name }}
-    </li>
-  </mat-tree-node>
-  <mat-nested-tree-node *matTreeNodeDef="let node; when: hasChild">
-    <li>
-      <div class="mat-tree-node">
-        <button
-          mat-icon-button
-          matTreeNodeToggle
-          [attr.aria-label]="'toggle ' + node.name"
-        >
-          <mat-icon class="mat-icon-rtl-mirror">
-            {{ treeControl.isExpanded(node) ? "expand_more" : "chevron_right" }}
-          </mat-icon>
-        </button>
-        {{ node.name }}
-      </div>
-      <ul [class.example-tree-invisible]="!treeControl.isExpanded(node)">
-        <ng-container matTreeNodeOutlet></ng-container>
-      </ul>
-    </li>
-  </mat-nested-tree-node>
-</mat-tree>
 
 <h3>Table</h3>
 <div class="mat-elevation-z8">
