+++ src/app/app.component.scss
@@ -38,17 +38,6 @@ h3 {
   margin: 64px 0 0;
 }
 
-.example-tree-invisible {
-  display: none;
-}
-
-.example-tree ul,
-.example-tree li {
-  margin-top: 0;
-  margin-bottom: 0;
-  list-style-type: none;
-}
-
 table {
   width: 100%;
 }
