+++ src/app/input/input.component.html
@@ -1,4 +1,4 @@
-<h1>Inputs</h1>
+<h1>Material Forms Inputs</h1>
 
 <mat-grid-list cols="2" rowHeight="2:1">
   <mat-grid-tile>
