+++ src/app/tab/tab.component.scss
@@ -1,0 +1,20 @@
+form {
+  display: flex;
+  flex-direction: column;
+  width: 50%;
+}
+
+.example-tree-invisible {
+  display: none;
+}
+
+.example-tree ul,
+.example-tree li {
+  margin-top: 0;
+  margin-bottom: 0;
+  list-style-type: none;
+}
+
+mat-tree {
+    background-color: transparent;
+}
