+++ src/app/tab/tab.component.ts
@@ -1,0 +1,45 @@
+import { NestedTreeControl } from "@angular/cdk/tree";
+import { Component } from "@angular/core";
+import { MatTreeNestedDataSource } from "@angular/material/tree";
+
+@Component({
+  selector: "tab",
+  templateUrl: "./tab.component.html",
+  styleUrls: ["./tab.component.scss"]
+})
+export class TabComponent {
+  dataSource = new MatTreeNestedDataSource<FoodNode>();
+  treeControl = new NestedTreeControl<FoodNode>(node => node.children);
+
+  constructor() {
+    this.dataSource.data = TREE_DATA;
+  }
+
+  hasChild = (_: number, node: FoodNode) =>
+    !!node.children && node.children.length > 0;
+}
+
+interface FoodNode {
+  name: string;
+  children?: FoodNode[];
+}
+
+const TREE_DATA: FoodNode[] = [
+  {
+    name: "Fruit",
+    children: [{ name: "Apple" }, { name: "Banana" }, { name: "Fruit loops" }]
+  },
+  {
+    name: "Vegetables",
+    children: [
+      {
+        name: "Green",
+        children: [{ name: "Broccoli" }, { name: "Brussel sprouts" }]
+      },
+      {
+        name: "Orange",
+        children: [{ name: "Pumpkins" }, { name: "Carrots" }]
+      }
+    ]
+  }
+];
